Added nvim release script

+++ release_scripts/tandem_release.sh
@@ -5,9 +5,9 @@
 # =======================================
 set -e  # terminate script if any command errors
 
-USAGE="Usage:  ./tandem_release <vim|sublime> /path/to/repo"
+USAGE="Usage:  ./tandem_release <vim|nvim|sublime> /path/to/repo"
 
-if [[ "$1" != "vim" ]] && [[ "$1" != "sublime" ]]; then
+if ! [[ "$1" =~ ^(vim|nvim|sublime)$ ]]; then
   echo $USAGE
   echo "ERROR: Please supply a valid plugin type."
   exit 1
@@ -39,6 +39,7 @@ TARGET_REPOSITORY_PATH="$2"
 # =======================================
 cd ./prepare_scripts
 ./$PLUGIN_TYPE.sh "../$TARGET_REPOSITORY_PATH"
+echo "Plugin files copied and prepared to commit."
 
 # =======================================
 #                Commit
@@ -48,14 +49,24 @@ MONOREPO_HASH=$( git rev-parse master )
 
 cd $TARGET_REPOSITORY_PATH
 git add .
-git commit -m "Cut release from $MONOREPO_HASH" --author="Team Lightly <teamlightly@gmail.com>"
-git push origin master  # Repository should have the main remote set to "origin"
-
-echo "Release $MONOREPO_HASH authored and pushed."
+CHANGED=$(git diff-index --name-only HEAD --)
+if [ -n "$CHANGED" ]; then
+  git commit -m "Cut release from $MONOREPO_HASH" --author="Team Lightly <teamlightly@gmail.com>"
+  git push origin master  # Repository should have the main remote set to "origin"
+  echo "Release $MONOREPO_HASH authored and pushed."
+else
+  echo "There were no changes to the repository and nothing was committed."
+fi
 
 # =======================================
 #                Release
 # =======================================
+echo -n "Are you sure you want to continue releasing (y/n)? "
+read answer
+if [[ $answer != "y" ]] && [[ $answer != "Y" ]] ;then
+  exit 1
+fi
+
 cd $RELEASE_SCRIPT_PATH
 pip install pygithub &
 pip install semver &
